agent/infra-agent/requirements.txt
psutil==5.9.7
httpx==0.26.0
python-dotenv==1.0.0
